Split out macro
Split out the macro into its own file.  Create a lib.rs file and rename
main.rs
