+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "Rust-testing"
+name = "rust_testing"
 version = "0.1.0"
 edition = "2021"
